arthimetic_operations.py
       
'arthimetic_operations.py'
# This class does arithmetic operations
class ArthimeticClass:
    def add(x, y):
        # adding two numbers
        if not isinstance(x, (int, float)) or not isinstance(y, (int, float)):
            raise TypeError("Both inputs should be numbers")
        return x + y
    def subtract(x, y):
        # subtracting two numbers
        return x - y
    def multiply(x, y):
        # multiplying two numbers
        return x * y
    def divide(x, y):
        # dividing two numbers
        return x / y